Revert "SPI encoder chain of 3 ported into devel branch."
This reverts commit 619e0b664d5347507c93cff76aaded74f11c2161.

+++ emBODY/eBcode/arch-arm/embobj/plus/mc/AbsEncoder.c
@@ -97,7 +97,7 @@ void AbsEncoder_config(AbsEncoder* o, uint8_t ID/*, eOmc_EncoderType_t type*/, i
 //        return;
 //    }
     
-    AbsEncoder_config_resolution(o, resolution);
+    AbsEncoder_config_resolution(o, (float)resolution);
     
     if (!o->fake)
     {
@@ -112,7 +112,7 @@ void AbsEncoder_config(AbsEncoder* o, uint8_t ID/*, eOmc_EncoderType_t type*/, i
     o->state.bits.not_configured = FALSE;
 }
 
-void AbsEncoder_config_resolution(AbsEncoder* o, int32_t resolution)
+void AbsEncoder_config_resolution(AbsEncoder* o, float resolution)
 {
     if (!o->fake)
     {
